--- platforms/android/example/src/main/java/io/element/android/wysiwyg/poc/MentionType.kt
@@ -1,5 +1,0 @@
-package io.element.android.wysiwyg.poc
-
-enum class  MentionType {
-    User, Room
-}
