+++ platforms/android/example/src/main/java/io/element/android/wysiwyg/poc/matrix/Mention.kt
@@ -1,4 +1,4 @@
-package io.element.android.wysiwyg.poc
+package io.element.android.wysiwyg.poc.matrix
 
 /**
  * Utility model class for the sample app to represent a mention to a
@@ -10,6 +10,7 @@ sealed class Mention(
     abstract val key: String
     abstract val mentionType: MentionType
     val link get() = "https://matrix.to/#/$key$display:matrix.org"
+    val text get() = "$key$display"
 
     class Room(
         display: String
