+++ platforms/android/example/src/main/java/io/element/android/wysiwyg/poc/matrix/MentionType.kt
@@ -1,0 +1,5 @@
+package io.element.android.wysiwyg.poc.matrix
+
+enum class  MentionType {
+    User, Room
+}
